472691: Use ItemPropertyDescriptor to show display names of features
This requires
org.eclipse.emf.parsley.ui.provider.FeatureLabelCaptionProvider.getLabel(Composite,
EObject, EStructuralFeature) to take the EObject instead of EClass.
This is used only in forms and dialogs where the EObject is available.
It is NOT used for the table's columns' headers, where no object is
available when building the columns (even before we were using the
FeatureCaptionProvider directly).

Change-Id: I368ac009570f100ef659bc457699375eb416b622
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=472691