dsl.standalone maven artifacts
This is the maven artifact that should be used to compile a parsley
module with Maven.

This way we avoid that the standard dsl project depends on
standalone.lib which seems to generate problems when running the build
with Indigo: the standalone.lib fake SWT classes are picked up when
running DSL tests, and this will break most of the tests that rely on
Java type system.

Change-Id: I1a44d9588e74c7bd2f51f57ea989e0cbbc310321